+++ packages/crypto/src/errors.ts
@@ -13,8 +13,9 @@ export class CryptoError extends Error {
             enumerable: false,
             value: this.constructor.name,
         });
-
-        Error.captureStackTrace(this, this.constructor);
+        if (Error.captureStackTrace) {
+            Error.captureStackTrace(this, this.constructor);
+        }
     }
 }